Implement Automated Process for Adding PostgreSQL Trigger and Notification
In this commit, an automated process has been added to streamline the creation of a trigger function and pg_notify notification in the PostgreSQL database. A Python script has been introduced to execute the necessary SQL statements to create the trigger function and associate it with a trigger on the payments table. This automation enhances the ease of integrating payment notifications into the database.